refactor(manager): rename addTrailingSlash
